# contracts/iot-ingestion.openapi.yaml
openapi: 3.1.0
info:
  title: FIT4110 Lab 05 - IoT Ingestion API
  version: 0.1.0
  contact:
    name: "API Support"
    email: "support@example.com"
  description: |
    Contract dùng cho Lab 05 Docker Compose readiness, kế thừa logic của Lab 03/04.
    Payload chính dùng các trường device_id, metric, value, unit và timestamp.
    Boundary minh hoạ cho temperature là -40 đến 80.
servers:
  - url: http://localhost:8000
    description: Local or Docker Compose API service
  - url: http://localhost:4010
    description: Prism mock server
tags:
  - name: system
    description: Health and runtime information
  - name: readings
    description: Sensor telemetry ingestion

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [system]
      operationId: getHealth
      summary: Kiểm tra trạng thái service
      description: Public health check dùng để kiểm tra container hoặc mock server đã sẵn sàng.
      security: []
      responses:
        "200":
          description: Service đang hoạt động
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                ok:
                  value:
                    status: ok
                    service: iot-ingestion
                    version: 0.1.0
                    db: ok
                    ai: ok

  /readings:
    post:
      tags: [readings]
      operationId: createReading
      summary: Gửi một bản ghi cảm biến từ thiết bị IoT
      description: |
        Endpoint nhận telemetry từ ESP32 hoặc gateway.
        Request phải có device_id, metric, value và timestamp.
        Trong lab này, ví dụ boundary dùng nhiệt độ hợp lệ từ -40 đến 80.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorReadingCreate"
            examples:
              valid_temperature:
                value:
                  device_id: ESP32-LAB-A01
                  metric: temperature
                  value: 31.5
                  unit: celsius
                  timestamp: "2026-05-13T08:30:00+07:00"
              boundary_high_temperature:
                value:
                  device_id: ESP32-LAB-A01
                  metric: temperature
                  value: 80
                  unit: celsius
                  timestamp: "2026-05-13T08:30:00+07:00"
      responses:
        "201":
          description: Đã ghi nhận bản đọc cảm biến
          headers:
            X-Warning:
              description: Warning trả về khi reading được chấp nhận nhưng có rủi ro, ví dụ nhiệt độ cao.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReadingCreated"
              examples:
                created:
                  value:
                    reading_id: R-20260513-0001
                    device_id: ESP32-LAB-A01
                    metric: temperature
                    accepted: true
                    created_at: "2026-05-13T08:30:01+07:00"
        "401":
          description: Thiếu hoặc sai token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                unauthorized:
                  value:
                    type: https://smart-campus.local/problems/unauthorized
                    title: Unauthorized
                    status: 401
                    detail: Missing or invalid bearer token
                    instance: /readings
        "422":
          description: Request không hợp lệ
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                validation_error:
                  value:
                    type: https://smart-campus.local/problems/validation-error
                    title: Validation error
                    status: 422
                    detail: body.value must be less than or equal to 80
                    instance: /readings
        "429":
          description: Thiết bị gửi dữ liệu quá nhiều trong thời gian ngắn
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /readings/latest:
    get:
      tags: [readings]
      operationId: getLatestReadings
      summary: Lấy các bản ghi cảm biến mới nhất
      description: Trả về danh sách bản ghi mới nhất, có thể lọc theo thiết bị và giới hạn số dòng.
      parameters:
        - in: query
          name: device_id
          required: false
          schema:
            type: string
          example: ESP32-LAB-A01
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          example: 5
      responses:
        "200":
          description: Danh sách bản ghi mới nhất
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SensorReading"
              examples:
                latest:
                  value:
                    items:
                      - reading_id: R-20260513-0001
                        device_id: ESP32-LAB-A01
                        metric: temperature
                        value: 31.5
                        unit: celsius
                        timestamp: "2026-05-13T08:30:00+07:00"
                        created_at: "2026-05-13T08:30:01+07:00"
        "401":
          description: Thiếu hoặc sai token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: Query parameter không hợp lệ
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /readings/{reading_id}:
    get:
      tags: [readings]
      operationId: getReadingById
      summary: Lấy chi tiết một bản ghi cảm biến
      description: API lấy chi tiết một bản ghi telemetry dựa trên reading_id định dạng chuẩn.
      parameters:
        - name: reading_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^R-[0-9]{8}-[0-9]{4}$"
          example: R-20260513-0001
      responses:
        "200":
          description: Chi tiết bản ghi cảm biến
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
              examples:
                reading:
                  value:
                    reading_id: R-20260513-0001
                    device_id: ESP32-LAB-A01
                    metric: temperature
                    value: 31.5
                    unit: celsius
                    timestamp: "2026-05-13T08:30:00+07:00"
                    created_at: "2026-05-13T08:30:01+07:00"
        "401":
          description: Thiếu hoặc sai token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Không tìm thấy bản ghi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                not_found:
                  value:
                    type: https://smart-campus.local/problems/not-found
                    title: Not Found
                    status: 404
                    detail: Reading R-20260513-9999 does not exist
                    instance: /readings/R-20260513-9999

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token

  schemas:
    SensorMetric:
      type: string
      enum: [temperature, humidity, motion, smoke]

    SensorUnit:
      type: string
      enum: [celsius, percent, boolean, ppm]

    SensorReadingCreate:
      type: object
      required: [device_id, metric, value, timestamp]
      properties:
        device_id:
          type: string
          minLength: 3
          example: ESP32-LAB-A01
        metric:
          $ref: "#/components/schemas/SensorMetric"
        value:
          type: number
          minimum: -40
          maximum: 80
          description: |
            Giá trị mẫu dùng cho boundary test trong Lab 03 và Lab 04.
            Với metric=temperature, miền hợp lệ minh họa là -40 đến 80.
          example: 31.5
        unit:
          $ref: "#/components/schemas/SensorUnit"
        timestamp:
          type: string
          format: date-time
          example: "2026-05-13T08:30:00+07:00"

    SensorReading:
      allOf:
        - $ref: "#/components/schemas/SensorReadingCreate"
        - type: object
          required: [reading_id, created_at]
          properties:
            reading_id:
              type: string
              pattern: "^R-[0-9]{8}-[0-9]{4}$"
              example: R-20260513-0001
            created_at:
              type: string
              format: date-time
              example: "2026-05-13T08:30:01+07:00"

    SensorReadingCreated:
      type: object
      required: [reading_id, device_id, metric, accepted, created_at]
      properties:
        reading_id:
          type: string
          pattern: "^R-[0-9]{8}-[0-9]{4}$"
          example: R-20260513-0001
        device_id:
          type: string
          example: ESP32-LAB-A01
        metric:
          $ref: "#/components/schemas/SensorMetric"
        accepted:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2026-05-13T08:30:01+07:00"

    HealthResponse:
      type: object
      required: [status, service, version, db, ai]
      properties:
        status:
          type: string
          enum: [ok]
          example: ok
        service:
          type: string
          example: iot-ingestion
        version:
          type: string
          example: 0.1.0
        db:
          type: string
          enum: [ok, error]
          example: ok
        ai:
          type: string
          enum: [ok, error]
          example: ok

    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          example: https://smart-campus.local/problems/validation-error
        title:
          type: string
          example: Validation error
        status:
          type: integer
          minimum: 400
          maximum: 599
          example: 422
        detail:
          type: string
          example: body.value must be less than or equal to 80
        instance:
          type: string
          example: /readings